Rename styled.js to react.js and allow components as selectors
