+++ test/extract/extract.test.js
@@ -5,7 +5,7 @@ import renderer from 'react-test-renderer'
 import { basename } from 'path'
 import { matcher, serializer } from '../../jest-utils'
 import { fragment, injectGlobal, css } from '../../src/index'
-import styled from '../../src/styled'
+import styled from '../../src/react'
 
 expect.addSnapshotSerializer(serializer)
 expect.extend(matcher)
